Use /bin/sh instead of bash
`bash` may not be available, and even when available, its path may not be `/bin`. For example, on BSD systems, `bash` will be installed as `/usr/local/bin/bash`.

Anyway, `bash` is not required here.

In addition, the shell can be replaced with `ripgrep` since nothing else depends on its output.